editUser route + securing all admin routs

+++ controllers/companiesController.js
@@ -1,24 +1,36 @@
 const Company = require('../models/company')
 
 exports.getCompanies = (req, res) => {
-    Company.find({}, (err, companies) => {
-        res.json(companies)
-    })
+    if (req.user.role === 'admin') {
+        Company.find({}, (err, companies) => {
+            res.json(companies)
+        })
+    } else {
+        return res.status(401).send('Unauthorized request')
+    }
 }
 
 exports.saveCompany = (req, res) => {
-    let company = new Company(req.body)
-    Company.findByIdAndUpdate(company._id, company, { upsert: true }, function (err, savedCompany) {
-        if (err) console.log(err)
-        else res.status(200).json(company)
-    })
+    if (req.user.role === 'admin') {
+        let company = new Company(req.body)
+        Company.findByIdAndUpdate(company._id, company, { upsert: true }, function (err, savedCompany) {
+            if (err) console.log(err)
+            else res.status(200).json(company)
+        })
+    } else {
+        return res.status(401).send('Unauthorized request')
+    }
 }
 
 exports.removeCompany = (req, res) => {
-    Company.deleteOne({ _id: req.params.companyId }, (err,company) => {
-        if (err) console.log(err)
-        else {
-            res.json(company)
-        }
-    })
+    if (req.user.role === 'admin') {
+        Company.deleteOne({ _id: req.params.companyId }, (err, company) => {
+            if (err) console.log(err)
+            else {
+                res.json(company)
+            }
+        })
+    } else {
+        return res.status(401).send('Unauthorized request')
+    }
 }
